fix(OMN-12191): move routing enums out of models
